refactor: split highlight machine into rule-per-file pipeline
Address PR #368 review:
- Move the machine into a directory: index.ts reduces the pool through an
  ordered RULES array (each rule fed the previous rule's output), with one
  rule per file, replacing the hand-composed passes.
- Add per-rule BDD tests (1/2/3-input coverage and noop checks per rule).
