+++ CLAUDE.md
@@ -83,7 +83,7 @@ The authoritative schema lives in `supabase/schema/` as declarative SQL files, o
 ## Code conventions
 
 - **Models** (`app/models/`) hold domain types and pure transformation logic — no I/O.
-  - Session highlights are plain-data objects (`app/models/session-highlights.ts`): `derive*` functions produce a `SessionHighlight[]` pool, which `fetchSessionHighlights` passes through the three-pass machine in `app/models/highlight-refinement-machine.ts` (`remove → order → combine`). The plain data serialises across the server-action boundary; the client renders each via `renderHighlight` in `app/components/session-highlight-renderers.tsx` (a `type`-keyed renderer map, mapped-typed so every variant must have a renderer). Editorial refinements belong in a machine pass (removal/ordering/combining) or a renderer (rewording).
+  - Session highlights are plain-data objects (`app/models/session-highlights.ts`): `derive*` functions produce a `SessionHighlight[]` pool, which `fetchSessionHighlights` passes through the highlight machine in `app/models/highlight-refinement-machine/` — `index.ts` reduces the pool through an ordered `RULES` list (each rule fed the previous rule's output; ordering rules precede combining ones), one rule per file (`remove-*`, `order-*`, `combine-*`). The plain data serialises across the server-action boundary; the client renders each via `renderHighlight` in `app/components/session-highlight-renderers.tsx` (a `type`-keyed renderer map, mapped-typed so every variant must have a renderer). Editorial refinements belong in a machine rule (removal/ordering/combining) or a renderer (rewording); add a rule by writing a file and slotting it into `RULES`.
 - **Actions** (`app/actions/`) are `'use server'` functions that fetch data and return typed results.
 - **Components** (`app/components/`) and page files receive data as props; they do not fetch.
 - Route pages are in `app/(routes)/` — the `(routes)` group is just for organisation, it doesn't affect URLs.
